feat(#120): collapse analytics_manager + analytics_coordinator → analytics.py
Merges 3 source files into one consolidated analytics.py module:
- NoteAnalytics + NoteStats (from analytics.py)
- AnalyticsManager — pure-Python quality/orphan/stale metrics (from analytics_manager.py)
- AnalyticsCoordinator — link graph, age distribution, productivity (from analytics_coordinator.py)

analytics_manager.py and analytics_coordinator.py become one-line shims.

Also removes dead AI deps (AITagger, AISummarizer, AIConnections) from
NoteAnalytics.__init__ — they were instantiated but never called, and
violated the pure-Python import boundary defined in #119.

51 analytics tests pass (40 existing + 11 new contract tests).
